put tests in speech_dataset_processor folder and restructure it
