+++ tools/speech_dataset_processor/sdp/processors/asr_inference.py
@@ -16,7 +16,7 @@ import os
 import subprocess
 from pathlib import Path
 
-from processors.base_processor import BaseProcessor
+from sdp.processors.base_processor import BaseProcessor
 
 
 class ASRInference(BaseProcessor):
