+++ tools/speech_dataset_processor/tests/test_data_to_dropbool.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import pytest
-from tools.speech_dataset_processor.processors.modify_manifest.data_to_dropbool import (
+from sdp.processors.modify_manifest.data_to_dropbool import (
     DropASRErrorBeginningEnd,
     DropHighCER,
     DropHighLowCharrate,
